requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.34
pydantic==2.8.2
pydantic-settings==2.5.2
requests==2.32.3
httpx==0.27.0
feedparser==6.0.11
trafilatura==1.9.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
qdrant-client==1.9.1
langgraph==0.2.28
apscheduler==3.10.4
jinja2==3.1.4
markdown==3.6
openai==1.40.6
python-multipart==0.0.9
pytest==8.2.2
pytest-asyncio==0.23.8
pyyaml==6.0.2